requirements.txt
gym==0.26.2
matplotlib==3.5.2
notebook==6.4.12
numpy==1.23.1
pandas==1.4.3
PyYAML==6.0
seaborn==0.11.2
tqdm==4.64.0
